Working on Dashboard for List User component.
Trick for Admin user that no need to request employees DB many time when
they are loging into the system. But need to logout and in again
whenever they want to open the app at the beginning.

Next steps: Handle CRUD methods, not only update the DB but also update
the global state